% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yolov7_python.R
\name{ovml_yolo_detect}
\alias{ovml_yolo_detect}
\title{Detect objects in image using a YOLO network}
\usage{
ovml_yolo_detect(
  net,
  image_file,
  conf = 0.25,
  nms_conf = 0.45,
  classes,
  as,
  ...
)
}
\arguments{
\item{net}{yolo: as returned by \code{\link[=ovml_yolo]{ovml_yolo()}}}

\item{image_file}{character: path to one or more image files, or a single video file (mp4, m4v, or mov extension)}

\item{conf}{scalar: confidence level}

\item{nms_conf}{scalar: non-max suppression confidence level}

\item{classes}{character: vector of class names, only detections of these classes will be returned}

\item{as}{string: for object detection networks, "boxes" (default and only option); for pose detection "segments" (default) or "keypoints"}

\item{...}{: currently ignored}
}
\value{
A data.frame with columns "image_number", "image_file", "class", "score", "xmin", "xmax", "ymin", "ymax", "frame"
}
\description{
Works on a single input image only, at the moment.
}
\examples{
\dontrun{
  dn <- ovml_yolo()
  img <- ovml_example_image()
  res <- ovml_yolo_detect(dn, img)
  ovml_ggplot(img, res)
}
}
\seealso{
\code{\link[=ovml_yolo]{ovml_yolo()}}
}
